backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.51.0
requests==2.32.3
websockets==13.1
pydantic==2.9.2
bcrypt==4.2.0
